refactor(rs-sdk): move DPNS network tests from src/ to tests/, drop band-aid endpoint helper
Five #[cfg(test)] blocks in src/platform/dpns_usernames/{queries,contested_queries}.rs
were really network-integration tests — they only used pub methods on Sdk (search,
availability, resolve, contested-vote-state, etc.) and shipped their own band-aid
test_address.rs helper to read DASH_SDK_PLATFORM_* env vars. They belong in tests/,
where the existing fetch::Config harness already loads tests/.env via dotenvy/envy
and exposes address_list() against the very same env vars.

This commit:
- Relocates all five #[ignore] network tests into a new tests/dpns_usernames.rs
  integration binary that re-mounts fetch::config + fetch::generated_data
  (#[path] sub-modules) so it can call Config::new().address_list() instead of
  parsing a hardcoded URL.
- Folds the previously hardcoded tests/dpns_queries_test.rs (which also pinned
  https://52.12.176.90:1443) into the same file via the shared
  build_testnet_sdk() helper, deleting the old file.
- Deletes src/platform/dpns_usernames/test_address.rs (the band-aid) and the
  five #[cfg(test)] mod tests blocks. No public-API change — the production
  helpers all stay where they were.

Net: -test_address.rs, -2 inline mod tests, -dpns_queries_test.rs,
+1 unified tests/dpns_usernames.rs, +consistent .env-driven endpoint config.

Run with:
  cargo test -p dash-sdk --features generate-test-vectors -- --include-ignored
