+++ packages/rs-sdk/src/platform/dpns_usernames/mod.rs
@@ -1,7 +1,5 @@
 mod contested_queries;
 mod queries;
-#[cfg(test)]
-mod test_address;
 
 pub use contested_queries::ContestedDpnsUsername;
 pub use queries::DpnsUsername;
